Move definition of String from stdstring.c to stdstring_impl.h
